requirements.txt
oauthlib==0.7.2
pep8==1.6.2
pyflakes==0.8.1
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
wsgiref==0.1.2
